fixed crash with nitront

+++ aidanbot.py
@@ -150,10 +150,12 @@ class AidanBot(commands.Bot):
 		emogilesstext = re.split(r':\w*:(?!\d*>)', ctx.message.content)
 		if len(emogis) == 1 and emogilesstext[0] == "$" and emogilesstext[1] == "":
 			realemogi = get(self.emojis, name=emogis[0])
-			msgs = await ctx.channel.history(limit=2).flatten()
-			await msgs[1].add_reaction(realemogi)
-			await ctx.message.delete()
-			return True
+			if realemogi:
+				msgs = await ctx.channel.history(limit=2).flatten()
+				await msgs[1].add_reaction(realemogi)
+				await ctx.message.delete()
+				return True
+			return False
 		elif len(emogis) > 0:
 			txt = emogilesstext[0]
 			for idx, emogi in enumerate(emogis):
